requirements.txt
Flask
tensorflow
numpy
Pillow
opencv-python
matplotlib
pandas
scipy